Remove support for `fun_info(Fun, pid)`
No longer support retrieving the fun creator pid. An exception
will be raised if `erlang:fun_info/2` is called with `pid` as
the second argument, and `erlang:fun_info/1` will no longer include
the `{pid,Pid}` tuple.

It was mentioned as an upcoming potential incompatibility that we would do
this in Erlang/OTP 28:

https://www.erlang.org/doc/upcoming_incompatibilities.html#fun-creator-pid-will-be-removed