gh: Use data from reuse lint when scanning code
When scanning code for licenses, we now also include data
from `reuse lint` into the results. This allows us to have
more consistent results as reuse scanning is predictable in
a way that scancode is not.

We still keep scancode as it can find licenses in files that
do not yet use the `reuse` standard.