erts: Replace FP16 implementation with a new one
The new implementation has the same behavior as the previous one.
The reason we are writing this new implementation is that the
previous version is external and has an MIT license. If it gets a newer
version, we need to upgrade it and keep it up to date. Because FP16 has
a clear specification and is stand-alone, it is easier to just write
our own implementation.

The newer compilers already have native support for FP16, so this
implementation is only relevant for older compilers. For this reason,
the new implementation has not been tested for speed.