add extracted license information
update "hasExtractedLicensingInfos" to consider LicenseRefs that are not
included in the repo. this information has been hard-coded (the license
text) but there is a check to see that we only add the license if it is
ever present in the repo. this means that if this license
`LicenseRef-scancode-wxwindows-free-doc-3` is not present in any
curation nor SPDX license identifier, then it will not be added to the
resulting SPDX