[erts|esock] Send handle use for sendv
Since when sending an I/O vector (using sendv), the send may
be split into several sends, each using the same (send) handle.
So, to ensure that a OVERLAPPED structure of an direct send cannot
be mistaken for an SCHEDULED send, we must reset (set to undefined)
the send handle in the opeation for the direct send.

OTP-19482