[erts|esock] Handle partial (sendv) writes on Windows
Finalize the sendv for WIndows.
sendv was only partly implemented. The completion t hread part
was "missing".
The completion threads shall free (sendv) buffers and other
allocated data.
Fixed written type; Was signed instead of unsigned.

Since when sending an I/O vector (using sendv), the send may
be split into several sends, each using the same (send) handle.
So, to ensure that a OVERLAPPED structure of an direct send cannot
be mistaken for an SCHEDULED send, we must reset (set to undefined)
the send handle in the opeation for the direct send.

OTP-19482