Add missing wx dependencies
The following applications:

 - et
 - observer
 - debugger
 - reltool

And also the wx/examples/[simple,demo,sudoku] depends on wx application
in order to build properly.

When cross compiling Erlang/OTP [1], there are cases where the initial
bootstrap system (which in most of the cases is just the same Erlang/OTP
version but built for host) may not have Erlang wx application enabled.
Thus when cross compiling the above applications will not able to found
wx.hrl and wx behaviour files.

That is because the target build should look into ERL_TOP in order to
see the missing dependencies.

Without the proper ERL_COMPILE_FLAGS the build will try to use from the
host system, which in my case does not have wx enabled.

1: https://github.com/erlang/otp/blob/master/HOWTO/INSTALL-CROSS.md