Remove (wrong) non_empty_runq call
If we steal exactly one task, then we directly execute it, and it never
ends up in our runqueue. So it is wrong to update the flag telling
everyone that we have work they can steal.
If we steal more than one task, then we will call erts_add_runq_len,
which will take care of calling erts_non_empty_runq for us.