Use a queue rather than a stack for contended runqueues
There are two benefits:
- Being FIFO rather than LIFO, we are much more likely to find the
  runqueue uncontended on the second pass (since it is a runqueue that
  was found contended a lot earlier)
- We can now do several passes, rather than defaulting to a blocking
  call to lock on the second pass.

There was just one small issue: the implementation of equeues did not
support passing them across function boundaries, since it used a
preprocesor macro to find the default queue (when trying to grow the
queue). I fixed that by adding a field to the queue itself.