Use matrix to permute flavours and types in tests
Presently the "Build Erlang/OTP (Flavors and Types)" job in testing
github workflow uses a matrix to parallely test building the jit and emu
flavours, and a bash for loop to sequentially test each type in each
flavour.

This commit updates the github workflow use a single matrix to permute
and parallely test every flavour-test permutation, thereby speeding up
the job and pin-pointing the flavour-test pair on a failure.