Optimise prim_file in write mode
This adds two optimisations to `prim_file`.

For files in various write modes, it skips allocating the read buffer -
this saves a lot on allocations when writing a lot of files.

For various write operations, if file is opened in just write mode,
we don't need to call `seek` to reset position before writing,
we can just write, since we couldn't have read from the file,
there's nothing to synchronise. This saves one syscall per write.