compiler: Fix alias analysis bug
Previously all bifs selected by the erl_internal:comp_op/2 predicate
was erroneously considered safe from an aliasing perspective. This is
not true as they, depending on the argument types, could access
internal elements of a term which has been destructively updated. As
this makes the destructive update visible, this is illegal and should
not happen.

To fix this problem, we have to consider all comparison bifs as
aliasing operations, thus avoiding the problem described
above. Unfortunately this is a very pessimistic solution, and by
making use of type information we can avoid forced aliasing when it
can be determined that the comparison bif will be reduced to a simple
tag test.

This is less conservative fix than the one presented in #9833.

Co-authored-by: Isabell H. <isabell@erlang.org>
Closes #9813.