erts, kernel: Little-endian packet header lengths
Add support for little-endian packet header lengths in gen_tcp and
erlang:decode_packet, for lengths of two, three and four.

Add support for packet header lengths of three, both LE and BE.

The packet option can now be a tuple {Len, Endianness}, where Len can
be 2, 3 or 4 and Endianness can be big or little. If endianness is big
the behaviour is the same as the old options.

Requested in issue #7265.