Optimize single generators in comprehensions
The only way to bind variables in a comprehension is via a generator.
If a value needs to be calculated once and used both as a filter and
as a result, one must resort to tricks such as:

    [V || E <- L, V -> [expensive_calculation(E)], some_filter(V)]

This commit optimizes comprehensions where a generator iterates over a
list known to contain a single element.