kernel: Add option for erl_boot_server listen port
Implements start/2 and start_link/2 for erl_boot_server, which accept
a map with configuration options.

The map key listen_port allows the configuration of the boot server
listening port. If an empty map is provided, or the listen_port is
zero, the old behavior is kept, i.e. an ephemeral port is used.

Requested in issue #7365.