Fix lock order violation
The helper function `suspended_proc_lock()` in erl_debugger.c could
cause a lock order violation dignostic in the lock checker in the
debug build. This commit fixes that problem, and also modifies some
of the tests cases to ensure that the lock checker will always catch
this problem.

While at it, use `spawn_link` instead of `spawn` in the test case to
ensure that processes are terminated at the end of each test case.