[kernel|esock] Handle 'select_sent'
The socket nif can actually return the atom 'select_sent'
when calling accept (prim_socket:accept). This is a race,
which is difficult to reproduce. But, this should be
translated into an {error, {invalid, select_sent}}.
This happens when there is already an ongoing accept,
which has just resulted in a select message having been sent
(select_sent).

OTP-19684