Fix stability in keysort/2
The optimization of leading equal elements was broken in two ways:

It used full term comparison as an optimization assuming that it
would be faster than looking at the elements, which could be
plain wrong if keysort is used for large tuples that are expensive
to compare but with small keys.

When transitioning to descending split in that the collected
element order had to be reversed, and the run of equal elements
had to be closed to ensure they all end up before the
descending sort elements.