Fix stability in sort/1
The comparisons for descending split were incorrect for
equal elements; should use `>`, not `>=`.  Maybe a copy paste error?
Keysort used the correct comparison and had apparently
gotten more attention.

Using mergel/2 after descending split was not entirely correct.
The correct variant does not exist.  For keysort/1 the possibility
to combine keymergel with argument desc exists, and this is the
correct combination; elements in sorted order in the sublists,
and sublists in reversed order (prefer the tail if equal).
This was solved by reversing the `Rs` list when calling mergel/2.

The optimization for leading equal terms was broken
when transitioning to descending split in the collected
element order had to be reversed, and the run of equal elements
had to be closed to ensure they all end up before the
descending sort elements.