Make export table size configurable
This is a change to the emulator that allows the export table size to be
configured at runtime.

The default size is 4000, which is the same as the initial size of the export
table.

I've set the maximum size to 2048*1024. Which I believe is a reasonable limit.
The default limit is 512*1024, which is the original hardcoded limit.

The reason i made this change is that in sufficiently large systems the export
list table can grow to exhaust the maximum indexes.