[erts] Increase number of connections epmd can handle on Windows
FD_SETSIZE increased from 64 to 1024.

The change in #10029 (OTP-19702) caused maximum amount of open connections
for epmd to be reached for some test cases connecting a large amount of nodes
at the same time. This since selection of dist module was parallelized
which caused more "port please" requests to be issued in parallel to epmd.