[snmp|test] Make sure manager and agent has the same (local host) info
Make sure that manager and agent use the same (local) address
info. Previously the address used by the agent and the manager
where retrieved at two different times. So if some network glich
occured, this  could cause the agent to use one address and the
manager get another for the agent (causing messages from the agent
to the manager to be rejected).