Eliminate the set_tuple_element instruction
The `set_tuple_element` instruction was used before OTP 26 to optimize chains of
`setelement/3` calls. For example:

    foo(S0) when tuple_size(S0) =:= 8 ->
        S1 = setelement(8, S0, a),
        S2 = setelement(7, S1, b),
        setelement(5, S2, c).

The compiler would keep the first call to `setelement/3` and replace
the other two with `set_tuple_element`, which would destructively
update the tuple created by the `setelement/3` call.

Starting in OTP 26, this was changed so that all three `setelement/3`
calls are replaced with a single `update_record` instruction.

The `set_tuple_element` instruction was only used when the size of the
tuple being updated was not known at compile time, as in the following
example:

    bar(S0) ->
        S1 = setelement(8, S0, a),
        S2 = setelement(7, S1, b),
        setelement(5, S2, c).

The only difference compared to the previous example is that the type
of `S0` is not known. When the size of the tuple cannot be determined
at compile time, the compiler cannot use `update_record` and must fall
back to `setelement/3` and `set_tuple_element`. However, the way this
fallback was implemented caused extra useless instructions to be
generated (see #10125 for examples).

This commit eliminates the use of `set_tuple_element` in the fallback
path, allowing us to remove some particularly messy code from
`beam_validator`.

Resolves #10125