Update remaining usages of app-less -include_lib()
`erl_tar.hrl` is a private header file of stdlib, but `erl_tar.erl` is including it with:

```
-include_lib("erl_tar.hrl").
```

While this works, it breaks some code-analysis tools that (too strictly?) expect an `-include_lib()` argument to be of the form `"APP/include/HEADER.hrl"`.

There doesn't seem to be any advantage to it this way, so let's use `-include("erl_tar.hrl").` instead.

In fact, this was almost the only case in otp where something was included this way. There were a couple of similar usages of `-include_lib()` in some dialyzer tests and in documentation, so we update those places here as well.