Improve pattern matching of binary segments
In the following example, the compiler would not understand
that the second clause could never succeed:

    decode(<<X:1/signed-unit:64>>) ->
        {int, X};
    decode(<<X:1/unsigned-float-unit:64>>) ->
        {float, X}.

In the following example:

    count_newlines(<<$\n,B/binary>>, Count) ->
        count_newlines(B, Count + 1);
    count_newlines(<<_:1/unit:8,B/binary>>, Count) ->
        count_newlines(B, Count);
    count_newlines(<<>>, Count) ->
        Count.

the size of the first segment is given in two different ways:
in the first clause it is given as size 8 and unit 1, and in
the second clause it is given as size 1 and unit 8.

The compiler would not understand that the size of segments were the
same and would generate suboptimal code for the second clause.

This commit improves the pattern matching compilation of binary
segments, which results in better code for the `count_newlines/2`
function, and the following warning for the `decode/1` function:

    t.erl:7:1: Warning: this clause cannot match because a previous clause at line 5 matches the same pattern as this clause
    %    7| decode(<<X:1/unsigned-float-unit:64>>) ->
    %     | ^