ssl: Correct connection state handling in TLS sender
Optimization commit 28f7e80233718f9f42fe8863370f7e99ee14d22c
broke max_fragment_length handling, that is the TLS sender
lost its knowledge of the maximum fragment length.

Make TLS sender process aware of it in the cases it is negotiated,
for default maximum we do not need to store it in the connection state.

Closes #10191