fix ignoring under_investigation
the ".github/scripts/otp-compliance.es vex run" was ignoring certain
OpenVEX statements if the CVE name existed. That meant an OpenVEX
statement could say "under_investigation" and the CVE reported a
vulnerability, and simply based on the CVE Id, the script would ignore
it as the CVE Id exists in OpenVEX.

This ticket fixes the script to filter CVE results that already exist in
OpenVEX statements.

OTP-19779