use fix up tests
add fixup otp_versions.table to get consistent results. The data is
fixed and the tests will be consistent. at the moment, each OTP
patch updates the otp_versions.table, and because the script uses
otp_versions.table as is, the OTP patch updates the otp_versions.table,
which was affecting the output of the tests.