backport otp-compliance and gh scripts
back ports from `master` branch otp-compliance and gh scripts.

otp-compliance backport is needed because some updates to the sbom were
incorportated for the `master` branch and they should have been added to
the `main` branch.

gh scripts are backported to `maint` branch so that PRs can scan for
vendor vulnerabilities. if a vendor vulnerability is found, it could be
because the PR updated the vendor libraries and there is already a
possible CVE on that version. in a bad case, it could just happen that
the PR contains a change in the vendor file and the same day that the PR
is created, some other unrelated vendor library gets a CVE. the script
will not distinguish between these two events. either way, a new GH
issue will be created for someone to look into it.

gh scripts are updated here as well in preparation for the adoption of
OpenVEX as VEX implementation.