add app-id as a secret
app-id is not a secret, but the reusable workflow would need to add yet
one more parameter for `uses` call in `maint.yml`, while the dispatch
workflow and scheduled job perform an API query and runs in a different
environment, so the app-id can be fetched as a `vars` from the repo
variable.

to make the integration consistent, we will just make the app-id private
and save passing one more argument.