fix vendor vulnerability for forked repos
on forked repos, the Github App will not used the secrets from the
Erlang/OTP repository. this means that anything that relies on the token
from the Erlang Scanner Bot App will not work. this is a GH security
measure, as otherwise, any PR towards Erlang/OTP from a forked repo
could trigger potential creation of issues, or any other permission.

because of that, the logic for running the reusable workflow in PRs has
to always set the `fail_if_cve` to `false`, as setting it to true tries
to trigger the creation of an issue in the Erlang/OTP repo.

the scheduled job is inside the Erlang/OTP, so the access to secrets is
allowed and can potentially create Issues, becomes it runs inside the
Erlang/OTP repo and it is considered thus, safe.