Unregister gen_server name in terminate
Currently, for every registered process there's a potential race condition
where when the process terminates, the supervisor can restart it before
the registry processes the DOWN message.
This isn't possible to avoid in all situations, but for servers that
trap_exit and process `terminate` calls, this is entirely possible.

This change implements this adding a call to `gen:unregister_name`
just after users `terminate` callback is processed.