wx: Fixed the types of pixel values.
The GetRed, GetBlue, GetGreen in wxWidgets return `unsigned char`,
But the code in wxe_wrapper_4.cpp is using `char` to receive the result
and sign extending it into a int.

As a result, a pixel whose red is `250`(`16#FA`) will be returned as
`4294967290`(`16#FFFFFFFA`) by `wxImage:getRed/3`.

This commit fixed this problem.