beam_ssa_type: Tighten bounds in the signatures pass
The signatures pass is run once, doing a type analysis of all functions
in the module to compute argument types and return types for all functions.

This pass must be conservative regarding the range for certain
operators. Consider this example:

    fact(N) when is_integer(N, 0, 10_000) ->
        fact(N, 1).

    fact(0, P) -> P;
    fact(N, P) -> fact(N - 1, P * N).

To ensure convergence, the signatures pass will compute the range of
the result of `N - 1` to be '-inf' to 9999, and `P * N` to be 1 to
'+inf'.

This commit improves the signatures pass by attempting to use a
tighter range for the result of the `+`, `-`, `*`, and
`bnot` operators. If the value range analysis converges, the tighter
range is kept; otherwise, the more conservative range is used.

For the example, the range for `N` will now be 0 to 10000.

Solves #8157