Make make.erl support most file types supported by erlc
This extension makes it possible to use `make` to compile almost
all source file types supported by `erlc`. The exception is `idl`
files, which have somewhat complex output.

The change is backwards-compatible, such that any time extensions
are omitted, `.erl` is assumed, and the same logic as before is used.
If the extension is given (and other than `.erl`), the work is passed
to `erl_compile:compile(Args, CWD)`, where `Args` is given by the new
`make` option `{erlc, Args}`, and is a string, such as would have been
given to the `erlc` command. Modification timestamps are compared as
for `.erl`/`.beam` pairs, but using the list of supported file types
in `erl_compile:extensions()`.

Furthermore, a new load option, `autoload` is introduced. This is the
default when running `erl -make`, and allows each `Emakefile` line
to specify a load option (`load` or `netload`). This allows for
parse-transform modules to be compiled and loaded selectively for use
by later steps in the `Emakefile`.