Fix `proc_lib` documentation of spawn/1 and similar functions
The documentation mentioned an equivalence of `erlang:apply(Fun)` but
this should read `erlang:apply(Fun, [])`.

This is not a strict equivalence as spawn/1 and similar functions do not
exactly set `initial_call/1` as if spawn/3 and similar functions would have
been called with `erlang, apply, [Fun, []]`, but the actual output of
`initial_call/1` is not entirely specified.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>