stdlib: Add classifying functions
When making scanners there might be a need to classify codepoints if
the specification follows the Unicode standard.

The following:
 is_whitespace/1,
 is_ID_start/1,
 is_ID_continue/1,
 category/1
is (at least) needed to scan and parse for example markdown.

This commit is backward incompatible if the *undocumented*
unicode_util:[is_]whitespace(..) functions have been used.

unicode_util:whitespace() have been changed to
unicode_util:pattern_whitespace() string module uses it and documents
that is uses the `pattern_whitespace` property.

So the name of the function have been changed, this is because
`pattern_whitespace` is not a subset of `whitespace` characters,
as was discovered during testing.

is_whitespace(Char) have been changed to use the Unicode `whitespace`
property.

See Unicode Standard Annex #31 and #44.