kernel: gen_sctp: document the {sctp_error, ...} message
The `{sctp_error, ...}` message is not documented, but it does exist.
The logic that generates both `{sctp, ...}` and `{sctp_error, ...}`
messages resides in inet_drv.c, specifically in packet_binary_message()
and sctp_parse_async_event().  Within sctp_parse_async_event(), an error
is indicated by replacing the initial `sctp` atom with `sctp_error`.

The following SCTP events are currently reported as errors:

* SCTP_SEND_FAILED (becomes #sctp_send_failed{}),
* SCTP_REMOTE_ERROR (becomes #sctp_remote_error{}),
* SCTP_PARTIAL_DELIVERY_EVENT (becomes #sctp_pdapi_event{}).

Document this and update the example code to print errors.