New shell functions `pi/1` and `pi/3` instead of `i/3`
Since the shell accepts the `<X.Y.Z>` syntax for pids, it was
surprising that `i(X,Y,Z)` worked, but `i(<X.Y.Z>)` did not, so
you could not simply paste a pid. The name `i/1` was however in
use by an undocumented function, and furthermore, the `i/0` info
function produces a very different output from `i/3`, so we instead
introduce the new names `pi/1` and `pi/3`, deprecating `i/3`.

Example:
    1> pi(<0.90.0>).
    [{current_function,{c,pinfo,1}},
     {initial_call,{erlang,apply,2}},
     {status,running},
    ...