Make the compiler warn about alias patterns that unify constructors
It is currently allowed to write a pattern such as `{a,B}={Y,Z} -> ...`
where both sides of the `=` are constructors. To track the variable
bindings, the compiler and other tools then need to be able to unify the
two sides, which may be arbitrarily complex. However, this feature is
practically never used in real code, and it is an unnecessary encumberance
to have to support it in all tooling.