Simplify grammar of map/record update expressions
It has always been allowed to chain map updates as well as record updates,
but the way this was expressed in the grammar was hardcoded and
restrictive, and it was left unmodified when the yecc grammar was
refactored to use precedence levels. This change simplifies the grammar and
allows map updates and record updates to be chained according to their
assigned precedence.