stdlib: Correct the pretty-printing of `catch Expr`
In 3171f2b7 and 762a18aa the precedence levels used by `erl_pp` to
format `catch Expr` was adjusted to avoid parentheses in many cases.
However, this only works in situations where the catch is on the right
of the surrounding operator, as in `Pat = catch Expr`. If the catch is
on the left, as in `(catch Expr) + 1`, the parentheses must be
preserved. The `erl_pp` prettyprinter is not clever enough to
distinguish between these situatuions, so for now we must accept some
unnecessary parentheses.

This restores the precedence level for printing and adds test cases to
validate the parsing and semantics as well as the pretty printing.