jit: Suppress i_mul_add optimization when adding squares
The `i_mul_add` instruction did not handle `X * X + X * X`
correctly because *both* source registers for the addition was the
same as the destination register for `X * X`. While the LHS was
correctly set to `X * X`, the RHS was read before the computation
of `X * X`, reading a stale value that could result in anything.

Fixes #10454