compiler: simplify v3_core
simplification of v3_core. the changes make application and function
calls to use ANF format. the only visible effect is the removal one
specific beam optimisation that deals with a case of cases, that is

```
case (case X of ... end) of ... end
```

this optimisation does not happen too often and the generated code is
almost equivalent in efficiency.