sbom: fixes sbom cache duplication
in the SBOM GH cache, some scan results have identical scanning
provenance and scanning options, yet they were executed at different
starting times and have different scanned files. as such, tools such as
oss-review-toolkit cannot differentiate which of the scan results with
the same provenance and scanning option is the correct one.

this commit merges scan results that contain the same scanning options
and provenance. the scan results are guaranteed to contain disjoint
files due to a previous deduplication pass, so the result of scan
results with same scanning options and provenance is simply the merge of
their licenses and copyrights.